% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsvbR.R
\name{biomass_components}
\alias{biomass_components}
\title{biomass_components() : compute biomass components in pounds.}
\usage{
biomass_components(plot, tree, fia_spp, division, vtotib, dbh, height)
}
\arguments{
\item{plot}{: integer | plot number}

\item{tree}{: integer | tree number}

\item{fia_spp}{: integer | FIA species code}

\item{division}{: string  | FIA ecological division}

\item{vtotib}{: double  | total inside bark volume of tree (top and stump) (cubic feet)}

\item{dbh}{: double  | diameter inside bark (inches)}

\item{height}{: double  | total height (feet)}
}
\value{
Returns a data.frame with the biomass components for each tree in following variables:
\itemize{
    \item plot : plot number
    \item tree : tree number
    \item wood : weight of biomass in wood 
    \item bark : weight of biomass in bark
    \item branch : weight of biomass in branches
    \item foliage : weight of biomass in foliage
    \item total : total weight of all components except foliage
    \item above_ground_biomass : total weight of above ground biomass including foliage
}
}
\description{
Compute biomass components in pounds using the National Scale Volume and Biomass estimators (NSVB).
}
\note{
\itemize{
    \item Returns 0.0 pounds for those species defaulting to the Jenkins equation that are woodland species (code 10).
    \item Recognized ecological divisions: "blank", 130, 210, 220, 230, 240, 260, 340, M210, M220, M230, M240, M260, M310, M330.
}
}
\examples{
# compute biomass components for a Douglas-fir tree with 10 inch dbh and 60 foot height in Ecological division M240
v_ib <- volib( 1, 1, 202, "M240", 10, 60 )
v_ob <- volob( 1, 1, 202, "M240", 10, 60 )
biomass_components( 1, 1, 202, "M240", v_ib$volib, 10, 60 )

}
