% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsvbR.R
\name{green_tons}
\alias{green_tons}
\title{green_tons() : computes green tons of a log segment or tree.}
\usage{
green_tons(plot, tree, fia_spp, vtotob, vtotib)
}
\arguments{
\item{plot}{: integer | plot number}

\item{tree}{: integer | tree number}

\item{fia_spp}{: integer | FIA species code}

\item{vtotob}{: double  | total outside bark volume of tree or log (cubic feet)}

\item{vtotib}{: double  | total inside bark volume of tree or log (cubic feet)}

\item{division}{: string  | FIA ecological division}
}
\value{
Returns a data.frame with the green tons for each tree or log in following variables:
\itemize{
    \item plot : plot number
    \item tree : tree number
    \item green_tons : green tons in the tree or log
}
}
\description{
Compute green tons using  Miles, et al. 2009.
}
\examples{
# compute green tons for a Douglas-fir tree with 10 inch dbh and 60 foot height 
v_ib <- volib( 1, 1, 201, "", 10, 60 )
v_ob <- volob( 1, 1, 201, "", 10, 60 )
green_tons( 1, 1, 201, v_ob$volob, v_ib$volib )

}
