% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsvbR.R
\docType{package}
\name{nsvbR-package}
\alias{nsvbR}
\alias{nsvbR-package}
\title{nsvbR}
\description{
Provides access to C++ implementation of the National Scale Volume and Biomass estimators (NSVB).
}
\details{
Developed by:
Greg Johnson
Greg Johnson Biometrics LLC
(c) October 2024
}
