% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsvbR.R
\name{volib}
\alias{volib}
\title{volib() : computes cubic foot volume inside bark of a tree}
\usage{
volib(plot, tree, fia_spp, division, dbh, height)
}
\arguments{
\item{plot}{: integer | plot number}

\item{tree}{: integer | tree number}

\item{fia_spp}{: integer | FIA species code}

\item{division}{: string  | FIA ecological division}

\item{dbh}{: double  | diameter inside bark (inches)}

\item{height}{: double  | total height (feet)}
}
\value{
Returns a data.frame with the cubic foot volume for each tree in following variables:
\itemize{
    \item plot : plot number
    \item tree : tree number
    \item volib : cubic foot volume inside bark
}
}
\description{
Compute cubic foot volume inside bark using the National Scale Volume and Biomass estimators (NSVB).
}
\examples{
# compute cubic foot volume inside bark for a Douglas-fir tree with 10 inch dbh and 60 foot height 
volib( 1, 1, 202, "M240", 10, 60 )

}
